@@ -44,21 +44,39 @@
     "babel-runtime": "^6.9.2",
     "bluebird": "^3.4.0",
     "body-parser": "^1.15.1",
+    "chai": "^3.5.0",
+    "codemirror": "^5.19.0",
     "ejs": "^2.4.2",
     "expect": "^1.20.2",
     "express": "^4.13.4",
     "express-jwt": "^3.4.0",
     "glob": "^7.0.3",
+    "history": "^2.1.2",
+    "immutable": "^3.8.1",
     "json-loader": "^0.5.4",
     "jsonwebtoken": "^7.1.7",
     "jwks-rsa": "^1.1.0",
+    "jwt-decode": "^2.0.1",
     "lodash": "^3.10.1",
     "lru-memoizer": "^1.5.0",
     "moment": "^2.13.0",
     "morgan": "^1.7.0",
     "ms": "^0.7.1",
     "nconf": "^0.8.4",
+    "nock": "^8.0.0",
+    "react": "^0.14.8",
+    "react-codemirror": "^0.2.6",
+    "react-dom": "^0.14.8",
+    "react-redux": "^4.4.5",
+    "react-router": "^2.6.1",
+    "react-router-redux": "^4.0.5",
+    "redux": "^3.5.2",
+    "redux-form": "^4.2.2",
+    "redux-logger": "^2.6.1",
+    "redux-thunk": "^2.1.0",
+    "redux-promise-middleware": "^2.4.0",
     "request": "^2.34.0",
+    "reselect": "^2.5.3",
     "safe-eval": "^0.3.0",
     "semver": "^5.1.0",
     "supertest": "^2.0.0",
@@ -71,27 +89,22 @@
     "babel-loader": "^6.2.4",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
-    "chai": "^3.5.0",
     "classnames": "^2.2.5",
     "cross-env": "^2.0.1",
     "css-loader": "^0.23.1",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-babel": "^3.2.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^5.2.2",
+    "eslint": "^3.3.1",
+    "eslint-config-auth0": "^6.0.1",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-import": "^1.13.0",
+    "eslint-plugin-jsx-a11y": "^2.1.0",
+    "eslint-plugin-react": "^6.1.2",
     "expect": "^1.20.2",
     "exports-loader": "^0.6.3",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
-    "history": "^2.1.2",
-    "immutable": "^3.8.1",
     "imports-loader": "^0.6.5",
     "json-loader": "^0.5.4",
-    "jwt-decode": "^2.0.1",
     "mocha": "^3.0.0",
-    "nock": "^8.0.0",
     "nodemon": "^1.9.2",
     "nyc": "^8.1.0",
     "postcss-focus": "^1.0.0",
@@ -99,29 +112,17 @@
     "postcss-reporter": "^1.3.3",
     "postcss-simple-vars": "^1.2.0",
     "raw-loader": "^0.5.1",
-    "react": "^0.14.8",
     "react-bootstrap": "^0.30.0",
-    "react-codemirror": "^0.2.6",
-    "react-dom": "^0.14.8",
     "react-loader-advanced": "^1.4.0",
     "react-pure-render": "^1.0.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.6.1",
-    "react-router-redux": "^4.0.5",
     "react-select": "^1.0.0-rc.1",
     "react-transform-hmr": "^1.0.4",
-    "redux": "^3.5.2",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
-    "redux-form": "^4.2.2",
-    "redux-logger": "^2.6.1",
-    "redux-promise-middleware": "^2.4.0",
     "redux-simple-router": "^2.0.4",
     "redux-static": "^1.0.0",
-    "redux-thunk": "^2.1.0",
     "request-promise": "^4.1.1",
-    "reselect": "^2.5.3",
     "rimraf": "^2.5.2",
     "style-loader": "^0.13.1",
     "supertest": "^2.0.0",
