@@ -53,13 +53,13 @@ export default class Editor extends Component {
   }
 
   componentDidMount() {
-    const { editor } = this.refs;
+    const { editor } = this.refs; // eslint-disable-line react/no-string-refs
     editor.getCodeMirror().refresh();
   }
 
   componentWillReceiveProps(nextProps) {
     if (nextProps.value && !this.state.value) {
-      const { editor } = this.refs;
+      const { editor } = this.refs; // eslint-disable-line react/no-string-refs
       if (editor) {
         editor.getCodeMirror().setValue(nextProps.value);
       }
@@ -67,7 +67,7 @@ export default class Editor extends Component {
   }
 
   componentDidUpdate() {
-    const { editor } = this.refs;
+    const { editor } = this.refs; // eslint-disable-line react/no-string-refs
     editor.getCodeMirror().refresh();
   }
 
@@ -85,11 +85,11 @@ export default class Editor extends Component {
     const { value, options } = this.props;
     return (
       <CodeMirror
-        ref="editor"
+        ref="editor" // eslint-disable-line react/no-string-refs
         value={this.state.value || value || ''}
         onChange={this.onChange}
         options={options}
       />
-  );
+    );
   }
 }
