@@ -14,7 +14,7 @@ const initialState = {
   user: null
 };
 
-export const auth = createReducer(fromJS(initialState), {
+export const auth = createReducer(fromJS(initialState), { // eslint-disable-line import/prefer-default-export
   [constants.LOGIN_PENDING]: (state) =>
     state.merge({
       ...initialState,
