@@ -3,7 +3,7 @@ import { fromJS } from 'immutable';
 import * as constants from '../constants';
 import createReducer from '../utils/createReducer';
 
-export const scripts = createReducer(fromJS({ }), {
+export const scripts = createReducer(fromJS({ }), { // eslint-disable-line import/prefer-default-export
   [constants.FETCH_SCRIPT_PENDING]: (state, action) =>
     state
       .setIn([ action.meta.name ], fromJS({ loading: true, error: null, script: null })),
