@@ -1,8 +1,9 @@
 import { Router as router } from 'express';
+import { middlewares } from 'auth0-extension-express-tools';
 
 import config from '../lib/config';
 import logger from '../lib/logger';
-import { middlewares } from 'auth0-extension-express-tools';
+
 
 export default () => {
   const hookValidator = middlewares
