@@ -1,9 +1,8 @@
 import expect from 'expect';
-import { users } from '../../../client/reducers/users';
-import * as constants from '../../../client/constants';
-
 import moment from 'moment';
 
+import { users } from '../../../client/reducers/users';
+import * as constants from '../../../client/constants';
 
 const initialState = {
   loading: false,
